# src/main/resources/openapi.yaml
openapi: 3.0.7
info:
  title: Shoe Retrieval Spring
  description: Retrieving shoe information from the blockchain
  version: 1.0.0
tags:
  - name: 'Product Information'
paths:
  '/recycables':
    get:
      description: Get products that can be recycled
      operationId: getRecycableProducts
      tags:
        - Recycables
      parameters:
        - name: ownerAddress
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProductInformation'

  '/recycables/{id}':
    post:
      description: Recycle a product
      operationId: recycleProduct
      tags:
        - Recycables
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecyclingRequest"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string


  '/products/{id}':
    get:
      description: Get product
      operationId: getProduct
      tags:
        - product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductInformation'
  '/products':
    post:
      description: Buy product
      operationId: buyProduct
      tags:
        - product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInformation"
      responses:
        200:
          description: OK
          content:
            application/jpeg:
              schema:
                type: string
                format: byte

    get:
      description: Get products
      operationId: getProducts
      tags:
        - product
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProductInformation'


components:
  schemas:
    ProductInformation:
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        owner:
          type: string
        ownerAdress:
          type: string
    Resource:
      type: object

    RecyclingRequest:
      properties:
        couponAmount:
            type: number
        toAddress:
            type: string